+++ CHANGELOG.md
@@ -11,6 +11,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ### Task
 * Eslint: Implementeer eslint-plugin-import ([#3065](https://github.com/dso-toolkit/dso-toolkit/issues/3065))
 
+### Deprecated
+* Tooltip: Deprecate ([#3163](https://github.com/dso-toolkit/dso-toolkit/issues/3163))
+
 ## ☀️ Release 73.1.0 - 2025-05-26
 
 ### Added